chore: install prettier and reformat main files

+++ src/commands/update.js
@@ -1,30 +1,30 @@
-const ProPublica = require('../propublica');
-const keys = require('../keys');
+const ProPublica = require('../propublica')
+const keys = require('../keys')
 
-exports.command = 'update';
+exports.command = 'update'
 
-exports.describe = 'refresh local cache of congressmembers';
+exports.describe = 'refresh local cache of congressmembers'
 
 exports.builder = {
-    force: {
-        describe: 'delete the local cache before updating',
-        type: 'boolean'
-    },
-    congress: {
-        describe: 'congress to use when pulling members',
-        type: 'number',
-        nargs: 1,
-        default: ProPublica.guessSession().congress,
-        requiresArg: true
-    }
-};
+  force: {
+    describe: 'delete the local cache before updating',
+    type: 'boolean',
+  },
+  congress: {
+    describe: 'congress to use when pulling members',
+    type: 'number',
+    nargs: 1,
+    default: ProPublica.guessSession().congress,
+    requiresArg: true,
+  },
+}
 
 exports.handler = async argv => {
-    const pp = new ProPublica({
-        key: await keys.get('propublica'),
-        congress: argv.congress
-    });
-    if (argv.force)
-        await pp.repsCache.delete();
-    return pp.updateMems({ aggressive: true });
-};
+  const pp = new ProPublica({
+    key: await keys.get('propublica'),
+    congress: argv.congress,
+  })
+  if (argv.force) await pp.repsCache.delete()
+  console.log('updating')
+  return pp.updateMems({ aggressive: true })
+}
